Far3bis sources

+++ trunk/Far3bis/unicode_far/panel.cpp
@@ -2319,7 +2319,7 @@ int Panel::SetPluginCommand(int Command,int Param1,void* Param2)
 			PanelInfo Info;
 			FileList *DestPanel = ((FileList*)this);
 			DestPanel->PluginGetPanelInfo(Info);
-			Result = (int)DestPanel->PluginGetPanelItem(Info.CurrentItem,(FarGetPluginPanelItem*)Param2);
+			Result = (int)DestPanel->PluginGetPanelItem(static_cast<int>(Info.CurrentItem),(FarGetPluginPanelItem*)Param2);
 			break;
 		}
 
@@ -2378,8 +2378,8 @@ int Panel::SetPluginCommand(int Command,int Param1,void* Param2)
 
 			if (Info)
 			{
-				CurFile=Info->CurrentItem;
-				CurTopFile=Info->TopPanelItem;
+				CurFile=static_cast<int>(Info->CurrentItem);
+				CurTopFile=static_cast<int>(Info->TopPanelItem);
 			}
 
 			// $ 12.05.2001 DJ перерисовываемся только в том случае, если мы - текущий фрейм
